ilk yükleme

+++ src/main/java/com/akif/lightlife/dto/response/YiyecekResponse.java
@@ -9,8 +9,8 @@ public class YiyecekResponse {
 
     private Long id;
     private String ad;
-    private int kalori;
-    private double karbonhidrat;
-    private double protein;
-    private double yag;
+    private Double kalori;
+    private Double karbonhidrat;
+    private Double protein;
+    private Double yag;
 }
