+++ src/main/java/com/akif/lightlife/entity/Yiyecek.java
@@ -17,8 +17,8 @@ public class Yiyecek {
 
     private String ad;
 
-    private int kalori;             
-    private double karbonhidrat;    
-    private double protein;         
-    private double yag;              
+    private Double kalori;             
+    private Double karbonhidrat;    
+    private Double protein;         
+    private Double yag;              
 }
